WE CAN MOD BREAK STATEMENTS
right now it just does a goto to the last statement of the switch scope,
but we're getting there, pretty soon we can construct switch statements.

also as per current cosmo, you need to note that the so-called constants
can be int, uint, long, and ulong, so work's cut out.