can handle structs, arrays, and some recursion
here's the idea, we know that we are going to walk through each
individual declaration within the function. so, if we run into one of
these declarations, we just save its pointer in the context structure,
and if we did not clean it up, we just check before the next declaration
to attempt a late cleanup.

when calling valid_subu_bounds in check_usage, the invariant we assume
is that all substitutions before list->start have been completed. all
the conditionals before that call are to ensure this invariant.