simplify to use ACTUALLY
we don't need to be able to substitute ALL the constants (although it
should be possible for most of them), we just need to account for those
that are most likely to be switch-cased or put as default values for
structs.

So we can't use SYMBOLIC anymore, instead we use a new macro called
ACTUALLY, and check for uses of that instead.

The main issue seems to be with declarations across a long swathe of
code or large expressions within macros, otherwise this is mostly okay.