fix errors by compiling viless
https://github.com/RaymiiOrg/viless

viless is a minimal vi clone that builds with cosmopolitan. It has a
switch case usage with the baud constants, and I thought the plugin
would be able to handle it.

Found a bug in the ifswitch plugin code -- if we have a switch statement
with cases but no default, then the plugin would crash unexpectedly. Now
there's a quick check for that.

Also I added the baud constants to tmpconst.h.