moved the integer replacement to its own function
the core of the plugin is actually integer replacement -- we have an
integer #define that we modified, and now we are using the parse tree
info to patch it correctly. for every mod/patch we need to do the same
checks, so I want to have this as a separate file.