handle switch statements without defaults
suppose the switch statement does not have a default, and none of the
cases match, then nothing in the switch should execute, ie we should
jump directly to the end of the switch statement (to the exit_label).
this commit adds that change and passes the test added in the previous
commit.